test(nestjs-tools-file-storage): refactor in separate suite for each provider
